+++ apps/studio/components/interfaces/TableGridEditor/index.ts
@@ -1,3 +1,2 @@
-export { default as TableGridEditor } from './TableGridEditor'
-export { default as SidePanelEditor } from './SidePanelEditor/SidePanelEditor'
 export { default as EmptyState } from './EmptyState'
+export { default as SidePanelEditor } from './SidePanelEditor/SidePanelEditor'
