+++ apps/studio/data/table-editor/table-editor-query.ts
@@ -49,6 +49,8 @@ export const useTableEditorQuery = <TData = TableEditorData>(
     {
       enabled:
         enabled && typeof projectRef !== 'undefined' && typeof id !== 'undefined' && !isNaN(id),
+      refetchOnWindowFocus: false,
+      refetchOnMount: false,
       staleTime: 5 * 60 * 1000, // 5 minutes
       ...options,
     }
