+++ apps/studio/state/tabs.ts
@@ -1,7 +1,7 @@
 import { ENTITY_TYPE } from 'data/entity-types/entity-type-constants'
 import { NextRouter } from 'next/router'
 import { ReactNode } from 'react'
-import { proxy, subscribe } from 'valtio'
+import { proxy, subscribe, useSnapshot } from 'valtio'
 import { addRecentItem } from './recent-items'
 
 /**
@@ -13,8 +13,27 @@ import { addRecentItem } from './recent-items'
  * So do not use this as reference for writing new valtio stores - refer to database-selector instead
  */
 
+// Object mapping editor types to their corresponding tab types
+export const editorEntityTypes = {
+  table: ['r', 'v', 'm', 'f', 'p'],
+  sql: ['sql'],
+}
+
 // Define the type of tabs available in the application
 export type TabType = ENTITY_TYPE | 'sql'
+
+type CreateTabIdParams = {
+  r: { id: number }
+  v: { id: number }
+  m: { id: number }
+  f: { id: number }
+  p: { id: number }
+  sql: { id: string }
+  schema: { schema: string }
+  view: never
+  function: never
+  new: never
+}
 
 export interface Tab {
   id: string // Unique identifier for the tab
@@ -60,7 +79,7 @@ const defaultState: TabsState = {
 export const tabsStore = proxy<TabsStateMap>({})
 
 // Helper to get/create state for a specific ref
-export const getTabsStore = (ref: string | undefined): TabsState => {
+export const getTabsStore = (ref?: string): TabsState => {
   if (!ref) return proxy(defaultState)
   if (!tabsStore[ref]) {
     const stored = localStorage.getItem(getStorageKey(ref))
@@ -85,6 +104,11 @@ export const getTabsStore = (ref: string | undefined): TabsState => {
   }
 
   return tabsStore[ref]
+}
+
+export const useTabsStore = (ref?: string) => {
+  const store = getTabsStore(ref)
+  return useSnapshot(store)
 }
 
 // Subscribe to changes for each ref and save to localStorage
@@ -349,19 +373,6 @@ export const handleTabDragEnd = (
   handleTabNavigation(ref, tabId, router)
 }
 
-type CreateTabIdParams = {
-  r: { id: number }
-  v: { id: number }
-  m: { id: number }
-  f: { id: number }
-  p: { id: number }
-  sql: { id: string }
-  schema: { schema: string }
-  view: never
-  function: never
-  new: never
-}
-
 // Function to create a unique tab ID based on type and parameters
 export function createTabId<T extends TabType>(type: T, params: CreateTabIdParams[T]): string {
   switch (type) {
@@ -380,12 +391,6 @@ export function createTabId<T extends TabType>(type: T, params: CreateTabIdParam
     default:
       return ''
   }
-}
-
-// Object mapping editor types to their corresponding tab types
-export const editorEntityTypes = {
-  table: ['r', 'v', 'm', 'f', 'p'],
-  sql: ['sql'],
 }
 
 // Function to remove tabs based on their editor type
